requirements.txt
# filepath: d:\Projects\Active_Projects\DULMS-Notifications-BOT\requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
selenium>=4.9.0
requests>=2.28.0
python-dotenv>=1.0.0
sse-starlette>=1.6.1
pillow>=9.5.0
pydantic>=1.10.7
email-validator>=2.0.0
python-multipart>=0.0.6
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0